clearLastActive()
