+++ ui/src/app/pages/signed-in/inactivity-monitor.spec.tsx
@@ -2,11 +2,9 @@ import * as React from 'react';
 import { mount } from 'enzyme';
 
 import { environment } from 'environments/environment';
-import {
-  INACTIVITY_CONFIG,
-  InactivityMonitor,
-} from 'app/pages/signed-in/inactivity-monitor';
+import { InactivityMonitor } from 'app/pages/signed-in/inactivity-monitor';
 import * as Authentication from 'app/utils/authentication';
+import { setLastActive } from 'app/utils/inactivity';
 import { authStore, notificationStore } from 'app/utils/stores';
 
 import { waitOneTickAndUpdate } from 'testing/react-test-helpers';
@@ -22,10 +20,7 @@ describe(InactivityMonitor.name, () => {
       isSignedIn: true,
     });
 
-    window.localStorage.setItem(
-      INACTIVITY_CONFIG.LOCAL_STORAGE_KEY_LAST_ACTIVE,
-      String(Date.now() - environment.inactivityTimeoutSeconds * 1000 - 1)
-    );
+    setLastActive(Date.now() - environment.inactivityTimeoutSeconds * 1000 - 1);
 
     expect(notificationStore.get()).toBeNull();
     const wrapper = mount(<InactivityMonitor />);
